\name{duplicated2}
\alias{duplicated2}

\title{
Extend Duplicated function
}

\description{
Extend Duplicated function. Return all duplicated values.
}

\usage{
duplicated2(x)
}

\arguments{
    \item{x}{["any"]: }
}

% \details{}

% \value{}

% \references{}
% \author{}
% \note{}
% \seealso{}

% \examples{}

\keyword{duplicated2}
