\name{extractExpr}
\alias{extractExpr}

\title{
Extract a pattern into a character string
}

\description{
Extract a pattern defined with a regular expression in order to extract from a character string.
}

\usage{
extractExpr(pattern, text)
}

\arguments{
    \item{pattern}{["character"]: A character string or a regular expression.}
    \item{text}{["character"]: String or object where to extract the pattern.}
}

% \details{}

% \value{}

% \references{}
% \author{}
% \note{}
% \seealso{}

% \examples{}

\keyword{extractExpr}
