\name{mail}
\alias{mail}

\title{
Mail allows to send e-mail through R under unix system
}

\description{
Mail allows to send e-mail through R under unix system
}

\usage{
mail(from, subject, to, body, attachment)
}

\arguments{
    \item{from}{["character"]: }
    \item{subject}{["character"]: }
    \item{to}{["character"]: }
    \item{body}{["character"]: }
    \item{attachment}{["character"]: }
}

% \details{}

% \value{}

% \references{}
% \author{}
% \note{}
% \seealso{}

% \examples{}

\keyword{mail}
