\name{maxCores}
\alias{maxCores}

\title{
maxCores for counting maximum number of cores
}

\description{
maxCores for counting maximum number of cores
}

\usage{
maxCores(mc.cores = detectCores())
}

\arguments{
    \item{mc.cores}{["numeric"]: }
}

% \details{}

% \value{}

% \references{}
% \author{}
% \note{}
% \seealso{}

% \examples{}

\keyword{maxCores}
