\name{multiplot}
\alias{multiplot}

\title{
multiplot
}

\description{
multiplot
}

\usage{
multiplot(..., plotlist = NULL, rows = 1, cols = 1, layout = NULL)
}

\arguments{
    \item{...}{["ggplot"]: }
    \item{plotlist}{["ggplot"]: }
    \item{cols}{["numeric"]: }
    \item{rows}{["numeric"]: }
    \item{layout}{["matrix"]: }
}

% \details{}

% \value{}

% \references{}
% \author{}
% \note{}
% \seealso{}

% \examples{}

\keyword{multiplot}
