\name{myScriptsMickael-internal}
\alias{colSd}
\alias{rowSd}
\alias{colSe}
\alias{rowSe}
\alias{is.pair}


\title{~ Internal: myScriptsMickael objects and methods ~}
\description{Internal: myScriptsMickael objects and methods.}
\details{
These are not intended to be called by the user.\cr
If you have a specific need or for more details on myScriptsMickael, feel free to contact the author.
}

% \seealso{}

\author{Mickael Canouil \email{mickael.canouil@good.ibl.fr}}

\keyword{internal}
