\name{myScriptsMickael-package}
\alias{myScriptsMickael-package}
\docType{package}
\alias{myScriptsMickael}

\title{~ my scripts (Mickael) ~}

\description{
My personnal scripts (utils)
}

\details{\tabular{ll}{
Package: \tab myScriptsMickael\cr
Title: \tab my scripts (Mickael)\cr
Author: \tab Mickael Canouil\cr
Maintainer: \tab Mickael Canouil\cr
License: \tab GPL (>= 2)\cr
Depends: \tab R (>= 2.15.0), methods, parallel\cr
URL: \tab http://www-good.ibl.fr/\cr
Encoding: \tab UTF-8\cr
}}

\seealso{
Functions : \code{\link{classS4Creator}}, \code{\link{duplicated2}}, \code{\link{GC}}, \code{\link{header}},
            \code{\link{lsos}}, \code{\link{mail}}, \code{\link{maxCores}}, \code{\link{mclapply2}}
}

\author{Mickael Canouil \email{mickael.canouil@good.ibl.fr}}

\keyword{package}
\keyword{utils}
\keyword{myScriptsMickael}