\name{readFast}
\alias{readFast}

\title{
readFast for Parrallel Read
}

\description{
readFast for Parrallel Read
}

\usage{
readFast(file, header = TRUE, mc.cores = 1,
         what = character(), sep = "\t", ...)
}

\arguments{
    \item{file}{["character"]: }
    \item{header}{["logical"]: }
    \item{mc.cores}{["numeric"]: }
    \item{what}{["function"]: }
    \item{sep}{["character"]: }
    \item{\dots}{["any"]: }
}

% \details{}

% \value{}

% \references{}
% \author{}
% \note{}
% \seealso{}

% \examples{}

\keyword{readFast}
