\name{myScriptsMickael-internal}
\alias{convert2Sample}
\alias{corSelect}
\alias{.disj2Table}
\alias{.dudi2.mix}
\alias{.imputeMean}
\alias{.makeFigure}
\alias{makeHeader}
\alias{.makeTable}
\alias{.multipleLM.mix}
\alias{.multipleLM}
\alias{.quali2Bin}
\alias{.removeVarNA}
\alias{.s.arrowClass}
\alias{.scatter.quali.}
\alias{.var2MuchMod}
\alias{moviesOrder}
\alias{qqplot.ggplot2}
\alias{s.class.ggplot2}
\alias{manhattan.ggplot2}
\alias{bw}
\alias{makeReport}


\title{~ Internal: myScriptsMickael objects and methods ~}
\description{Internal: myScriptsMickael objects and methods.}
\details{
These are not intended to be called by the user.\cr
If you have a specific need or for more details on myScriptsMickael, feel free to contact the author.
}

% \seealso{}

\author{Mickael Canouil \email{mickael.canouil@good.ibl.fr}}

\keyword{internal}
